nfcpy test/nfcpy test/requirements.txt
libusb1==1.6.4
ndeflib==0.3.1
nfcpy==0.13.4
pip==6.1.1
pyDes==2.0.1
pyserial==3.4
setuptools==15.0
